+++ CHANGELOG.md
@@ -3,6 +3,11 @@ CHANGELOG for SecurityAuthorization
 
 This changelog references the relevant changes (bug, security fixes and new features) done.
 
+* 2.0.0 (2016-03-10)
+
+  * [enhancement] #5 - Remove SecurityAuthorization.setAuthenticatedUser() and refactoring SecurityAuthorization.isGranted() with new optional parameter user.
+  * [enhancement] #6 - Remove client UI-Helper for Blaze.
+
 * 1.0.1 (2016-01-06)
 
   * Add the reference to GIT in the package.